Introduce a settings screen and add navigation from ui widget.
